Fix a bug with elements being dynamically added
If an element was added to the DOM and received focus before the MutationObserver callback ran, the previously focused element would be incorrect. To fix this, after updating the set of elements in the MutationObserver callback, check the currently focused element and prefer that if it’s in the set.